Introduced BaseRelationProcessor
